# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VendorFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List all vendors
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /procurement:
    get:
      operationId: listPurchaseOrders
      tags: [procurement]
      summary: List purchase orders
      responses:
        "200":
          description: List of purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [procurement]
      summary: Create a purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /procurement/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [procurement]
      summary: Get a purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePurchaseOrder
      tags: [procurement]
      summary: Update a purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
    delete:
      operationId: deletePurchaseOrder
      tags: [procurement]
      summary: Delete a purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /procurement/{id}/status:
    patch:
      operationId: updatePurchaseOrderStatus
      tags: [procurement]
      summary: Update purchase order status (approve/reject/submit)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    StatusUpdate:
      type: object
      properties:
        status:
          type: string
        rejectionReason:
          type: ["string", "null"]
      required:
        - status
    Vendor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        category:
          type: string
        rating:
          type: number
        totalSpent:
          type: number
        status:
          type: string
          enum: [active, inactive, suspended]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - phone
        - address
        - category
        - rating
        - totalSpent
        - status
        - createdAt
        - updatedAt
    VendorInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        category:
          type: string
        rating:
          type: number
        status:
          type: string
          enum: [active, inactive, suspended]
      required:
        - name
        - email
        - phone
        - address
        - category
        - rating
        - status
    VendorUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        category:
          type: string
        rating:
          type: number
        status:
          type: string
          enum: [active, inactive, suspended]
    PurchaseOrderItem:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
        category:
          type: string
      required:
        - id
        - description
        - quantity
        - unitPrice
        - total
        - category
    PurchaseOrderItemInput:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        category:
          type: string
      required:
        - description
        - quantity
        - unitPrice
        - category
    PurchaseOrder:
      type: object
      properties:
        id:
          type: integer
        poNumber:
          type: string
        vendorId:
          type: integer
        vendorName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItem"
        totalAmount:
          type: number
        status:
          type: string
          enum: [draft, submitted, approved, rejected, received, cancelled]
        notes:
          type: ["string", "null"]
        dueDate:
          type: string
        createdBy:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - poNumber
        - vendorId
        - vendorName
        - items
        - totalAmount
        - status
        - dueDate
        - createdBy
        - createdAt
        - updatedAt
    PurchaseOrderInput:
      type: object
      properties:
        vendorId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItemInput"
        notes:
          type: string
        dueDate:
          type: string
      required:
        - vendorId
        - items
        - dueDate
    PurchaseOrderUpdate:
      type: object
      properties:
        vendorId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItemInput"
        notes:
          type: string
        dueDate:
          type: string
...